requirements.txt
jupyterlab==3.4.2
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
pathlib==1.0.1
scikit-learn==1.1.1
scipy==1.8.0
sklearn==0.0
tqdm==4.64.0
deepdish==0.3.7
torch==1.12.1
pyro-api==0.1.2
pyro-ppl==1.8.1
uncertainty-calibration==0.0.9
